+++ apps/zhi-cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # zhi-cli
 
+## 1.5.20
+
+### Patch Changes
+
+- zhi-log@1.14.20
+
+## 1.5.19
+
+### Patch Changes
+
+- zhi-log@1.14.19
+
 ## 1.5.18
 
 ### Patch Changes