+++ libs/zhi-cmd/CHANGELOG.md
@@ -1,5 +1,25 @@
 # zhi-cmd
 
+## 0.6.3
+
+### Patch Changes
+
+- feat: add exec log
+
+## 0.6.2
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-device@2.10.0
+
+## 0.6.1
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-device@2.9.0
+
 ## 0.6.0
 
 ### Minor Changes