+++ libs/zhi-device/CHANGELOG.md
@@ -1,5 +1,17 @@
 # zhi-device
 
+## 2.10.0
+
+### Minor Changes
+
+- feat: change plugin path
+
+## 2.9.0
+
+### Minor Changes
+
+- feat: add getRequirePath
+
 ## 2.8.1
 
 ### Patch Changes