+++ libs/zhi-log/CHANGELOG.md
@@ -1,5 +1,19 @@
 # zhi-log
 
+## 1.14.20
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-device@2.10.0
+
+## 1.14.19
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-device@2.9.0
+
 ## 1.14.18
 
 ### Patch Changes