+++ libs/zhi-xmlrpc-middleware/CHANGELOG.md
@@ -1,5 +1,19 @@
 # zhi-xmlrpc-middleware
 
+## 0.6.9
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-device@2.10.0
+
+## 0.6.8
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-device@2.9.0
+
 ## 0.6.7
 
 ### Patch Changes